Complete backend

+++ back-end/src/main/java/com/henrygphp/swd392hospitalsystem/Repositories/PatientHistoryRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PatientHistoryRepo extends JpaRepository<PatientHistory, Integer> {
+public interface PatientHistoryRepo extends JpaRepository<PatientHistory, Long> {
 }
